requirements.txt
flask-cors==5.0.1
Flask==3.1.0
torch==2.5.1
torchvision==0.20.1
numpy==2.1.3
pillow==11.3.0
waitress==2.1.2
opencv-python==4.11.0.86
